refactor: move public-facing read() and detect() into io module
